+++ com.hangum.tadpole.commons.sql/src/com/hangum/tadpole/engine/query/dao/system/ledger/LedgerDAO.java
@@ -20,6 +20,7 @@ import java.sql.Timestamp;
  */
 public class LedgerDAO {
 	int seq;
+	int db_seq;
 	int user_seq;
 	String userID="";
 	String cp_seq="";
@@ -31,7 +32,22 @@ public class LedgerDAO {
 	
 	public LedgerDAO() {
 	}
-	
+
+	/**
+	 * @return the db_seq
+	 */
+	public int getDb_seq() {
+		return db_seq;
+	}
+
+
+	/**
+	 * @param db_seq the db_seq to set
+	 */
+	public void setDb_seq(int db_seq) {
+		this.db_seq = db_seq;
+	}
+
 
 	/**
 	 * @return the userID
