+++ com.hangum.tadpole.commons.sql/src/com/hangum/tadpole/engine/query/internal/system/Tadpole-System-Ledger.xml
@@ -35,7 +35,7 @@
 <select id="getMySQLListLedger" resultClass="com.hangum.tadpole.engine.query.dao.system.ledger.LedgerDAO"  parameterClass="java.util.Map">
 <![CDATA[
 	SELECT a.*, b.email AS userID
-	FROM ledger_head a,
+	FROM historyhub_head a,
 	 	tadpole_user b
 	WHERE a.user_seq = b.seq
 		AND b.email like #strUser#
@@ -51,19 +51,19 @@
 	template insert
  -->
 <insert id="insert_ledger" parameterClass="com.hangum.tadpole.engine.query.dao.system.ledger.LedgerDAO">
-	INSERT INTO ledger_head (user_seq, cp_seq, comment, request_query) 
-	VALUES (#user_seq#, #cp_seq#, #comment#, #request_query#);
+	INSERT INTO historyhub_head (db_seq, user_seq, cp_seq, comment, request_query) 
+	VALUES (#db_seq#, #user_seq#, #cp_seq#, #comment#, #request_query#);
 	
 	<selectKey resultClass="java.lang.Integer">
 		SELECT max(seq) seq
-		FROM ledger_head 
+		FROM historyhub_head 
 		WHERE user_seq = #user_seq# 
 	</selectKey>
 </insert>
 
 <!-- 사용처를 수정합니다. -->
 <update id="update_ledger_result" parameterClass="com.hangum.tadpole.engine.query.dao.system.ledger.LedgerDAO">
-	UPDATE ledger_head
+	UPDATE historyhub_head
 	SET 
 		 result_yn 		= #result_yn#
 		,result_msg 	= #result_msg#
