+++ com.hangum.tadpole.rdb.core/src/com/hangum/tadpole/rdb/core/viewers/connections/ManagerViewer.java
@@ -39,7 +39,6 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.Tree;
-import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.IEditorReference;
 import org.eclipse.ui.IWorkbenchActionConstants;
 import org.eclipse.ui.IWorkbenchPage;
@@ -126,14 +125,14 @@ public class ManagerViewer extends ViewPart {
 					// 리소스 가져온다.
 					addManagerResouceData(userDB, false);
 					
-					if(userDB.is_isUseEnable()) {
-						// 싱글 클릭일때 에디터에 오픈된 화면이 없으면 에디터 화면이 열리도록 수정.
-						IEditorPart editor = EditorUtils.findSQLEditor(userDB);
-						if(editor == null) {
-							QueryEditorAction qea = new QueryEditorAction();
-							qea.run(userDB);
-						}
-					}
+//					if(userDB.is_isUseEnable()) {
+//						// 싱글 클릭일때 에디터에 오픈된 화면이 없으면 에디터 화면이 열리도록 수정.
+//						IEditorPart editor = EditorUtils.findSQLEditor(userDB);
+//						if(editor == null) {
+//							QueryEditorAction qea = new QueryEditorAction();
+//							qea.run(userDB);
+//						}
+//					}
 					
 					// Rice lock icode change event
 					managerTV.refresh(userDB, true);
