+++ targetProject/target-tadpolea-1.7.6.target
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<?pde version="3.8"?><target name="target-tadpolea-1.7.6" sequenceNumber="321">
+<?pde version="3.8"?><target name="target-tadpolea-1.7.6" sequenceNumber="324">
 <locations>
 <location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="slicer" includeSource="true" type="InstallableUnit">
 <unit id="org.eclipse.rap.gef.feature.feature.group" version="0.1.0.20170612-1850"/>
